requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
streamlit>=1.38,<2.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
numpy>=1.26,<3.0
anthropic>=0.34,<1.0
sentence-transformers>=3.0,<4.0
google-cloud-bigquery>=3.25,<4.0
pyarrow>=17.0,<19.0
python-dotenv>=1.0,<2.0
pydantic>=2.8,<3.0
sqlalchemy>=2.0,<3.0
db-dtypes>=1.2,<2.0
httpx>=0.27,<1.0
joblib>=1.4,<2.0
pytest>=8.3,<9.0
